Removed seed with to many es - created spawnchild file
